+++ semcore/tag/src/Tag.jsx
@@ -59,8 +59,8 @@ class RootTag extends Component {
   }
 
   getCircleProps() {
-    const { size } = this.asProps;
-    return { size };
+    const { size, color, resolveColor } = this.asProps;
+    return { size, color, resolveColor };
   }
 
   getTextProps() {
@@ -81,7 +81,7 @@ class RootTag extends Component {
     this.setState({ focusable: 'container' });
   };
   getCloseProps() {
-    const { getI18nText } = this.asProps;
+    const { getI18nText, color, resolveColor } = this.asProps;
     const id = this.asProps.id || `igc-${this.asProps.uid}-tag`;
 
     return {
@@ -91,6 +91,8 @@ class RootTag extends Component {
       'aria-label': getI18nText('remove'),
       onMount: this.handleCloseMount,
       onUnmount: this.handleCloseUnmount,
+      color,
+      resolveColor,
     };
   }
 
@@ -147,11 +149,23 @@ class RootTag extends Component {
 class RootTagContainer extends Component {
   static displayName = 'TagContainer';
   static style = style;
-
-  static enhance = [i18nEnhance(localizedMessages), uniqueIDEnhancement()];
+  static enhance = [i18nEnhance(localizedMessages), uniqueIDEnhancement(), resolveColorEnhance()];
+  static defaultProps = {
+    color: 'gray-500',
+    theme: 'primary',
+  };
 
   getTagProps() {
-    const { size, theme, color, disabled, active, uid, id: outerId, interactive } = this.asProps;
+    const {
+      size,
+      theme,
+      color,
+      disabled,
+      uid,
+      id: outerId,
+      interactive,
+      resolveColor,
+    } = this.asProps;
     const id = outerId || `igc-${uid}-tag`;
 
     return {
@@ -160,14 +174,34 @@ class RootTagContainer extends Component {
       size,
       theme,
       color,
-      active,
       tag: interactive ? 'button' : undefined,
       interactive,
+      resolveColor,
     };
   }
 
+  getCircleProps() {
+    const { color, resolveColor, size } = this.asProps;
+    return { color, resolveColor, size };
+  }
+
+  getAddonProps() {
+    const { color, resolveColor } = this.asProps;
+
+    return { color, resolveColor };
+  }
+
   getCloseProps() {
-    const { size, theme, color, disabled, active, uid, id: outerId, getI18nText } = this.asProps;
+    const {
+      size,
+      theme,
+      color,
+      disabled,
+      uid,
+      id: outerId,
+      getI18nText,
+      resolveColor,
+    } = this.asProps;
     const id = outerId || `igc-${uid}-tag`;
 
     return {
@@ -175,10 +209,10 @@ class RootTagContainer extends Component {
       size,
       theme,
       color,
-      active,
       id: `${id}-clear`,
       'aria-labelledby': `${id}-clear ${id}-text`,
       'aria-label': getI18nText('remove'),
+      resolveColor,
     };
   }
 
@@ -243,6 +277,17 @@ class RootCloseTagContainer extends Component {
   }
 }
 
+function TagContainerCircle(props) {
+  const SAddon = Box;
+  const SCircle = Root;
+  const { styles, color, resolveColor } = props;
+  return sstyled(styles)(
+    <SAddon tag-color={resolveColor(color)}>
+      <SCircle render={Box} />
+    </SAddon>,
+  );
+}
+
 function Text(props) {
   const SText = Root;
   const { styles } = props;
@@ -285,14 +330,20 @@ function Close(props) {
 
 function Addon(props) {
   const SAddon = Root;
-  const { styles } = props;
-  return sstyled(styles)(<SAddon render={Box} tag='span' />);
+  const { styles, color, resolveColor } = props;
+
+  const tagColor = React.useMemo(() => {
+    if (typeof resolveColor !== 'function') return;
+    return resolveColor(color);
+  }, [color, resolveColor]);
+
+  return sstyled(styles)(<SAddon render={Box} tag='div' tag-color={tagColor} />);
 }
 
 function Circle(props) {
   const SCircle = Root;
-  const { styles } = props;
-  return sstyled(styles)(<SCircle render={Box} tag='span' />);
+  const { styles, color, resolveColor } = props;
+  return sstyled(styles)(<SCircle render={Box} tag='span' tag-color={resolveColor(color)} />);
 }
 
 const Tag = createComponent(RootTag, {
@@ -308,7 +359,9 @@ const CloseTagContainer = createComponent(RootCloseTagContainer, {
 
 export const TagContainer = createComponent(RootTagContainer, {
   Tag,
+  Addon,
   Close: CloseTagContainer,
+  Circle: TagContainerCircle,
 });
 
 export default Tag;